+++ src/Controllers/Partial/FooterController.php
@@ -1,4 +1,7 @@
 <?php
+
+namespace Partial;
+
 class FooterController extends \Controller {
 	/**
 	 * Default entry point for footer partial
@@ -10,12 +13,12 @@ class FooterController extends \Controller {
 
 		$this->response->data["footer"] = $this->template->getViewPath("footer");
 
-		$this->response->data["stylesheets"] = array_merge(
+		$this->response->data["stylesheets"] = \Arr::merge(
 			$this->response->data["stylesheets"] ?? [],
 			$this->template->assets->getStylesheets("footer")
 		);
 
-		$this->response->data["javascript"] = array_merge(
+		$this->response->data["javascript"] = \Arr::merge(
 			$this->response->data["javascript"] ?? [],
 			$this->template->assets->getJavascript("footer")
 		);