+++ src/Libraries/Core/StatusCode/NotFound.php
@@ -2,16 +2,23 @@
 
 namespace Core\StatusCode;
 
-use \Core\StatusCode\StatusCode;
-
 /**
  * Throw this whenever you want to redirect the current controller to a not found controller
  */
-class NotFound extends StatusCode {
+class NotFound extends \Exception implements StatusCode {
 	/**
 	 * @return int
 	 */
 	public static function getHttpCode(): int {
 		return 404;
+	}
+
+	/**
+	 * Returns a class name matching the name
+	 * of the http error exception being thrown.
+	 * @return \Core\ClassName
+	 */
+	public static function getClassName(): \Core\ClassName {
+		return new \Core\ClassName("StatusCode\NotFound");
 	}
 }